refactor: significantly reduced size of Docker image at the cost of a bit longer runtime.
The size reduction is achieved by performing the unpacking of the ground truth shortly before running the simulation instead of during the building of the image. This increases the runtime of the study by several seconds to minutes, because the unpacking is now confirmed at the start of the study.